requirements.txt
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
streamlit==1.24.0
faiss-cpu==1.8.2
redis